packages/qa-agent/requirements.txt
browser-use>=0.1.40
langchain-anthropic>=0.3.0
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
